pyproject.toml
[tool.poetry]
name = "twitter-app"
version = "0.1.0"
description = ""
authors = ["Camilo Ramirez <camilo.ramirez@xing.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
numpy = "^1.22.3"
matplotlib = "^3.5.1"
spacy = "^3.3.0"
ipykernel = "^6.13.0"
scikit-learn = "^1.0.2"
tweepy = {extras = ["async"], version = "^4.10.0"}
fastapi = {extras = ["all"], version = "^0.79.0"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
